+++ tests/DateFormatter/DefaultFormatterTest.php
@@ -6,9 +6,8 @@ namespace Stadly\PasswordPolice\DateFormatter;
 
 use DateTime;
 use PHPUnit\Framework\TestCase;
-use Stadly\PasswordPolice\DateFormatter;
-use Stadly\PasswordPolice\WordFormatter;
-use Traversable;
+use Stadly\PasswordPolice\CharTree;
+use Stadly\PasswordPolice\Formatter;
 
 /**
  * @coversDefaultClass \Stadly\PasswordPolice\DateFormatter\DefaultFormatter
@@ -33,11 +32,10 @@ final class DefaultFormatterTest extends TestCase
             '11,07',
             '07, 11',
         ];
-
-        self::assertEquals($expected, array_unique(array_intersect(iterator_to_array($formatter->apply([
+        self::assertEquals($expected, array_intersect(iterator_to_array($formatter->apply([
             new DateTime('2001-02-03 07:11:19'),
             new DateTime('1907-11-19 21:13:58'),
-        ]), false), $expected)), '', 0, 10, true);
+        ]), false), $expected), '', 0, 10, true);
     }
 
     /**
@@ -47,12 +45,14 @@ final class DefaultFormatterTest extends TestCase
     {
         $formatter = new DefaultFormatter();
 
-        $next = $this->createMock(WordFormatter::class);
+        $next = $this->createMock(Formatter::class);
         $next->method('apply')->willReturnCallback(
-            static function (iterable $words): Traversable {
-                foreach ($words as $word) {
-                    yield strrev($word);
+            static function (CharTree $charTree): CharTree {
+                $charTrees = [];
+                foreach ($charTree as $string) {
+                    $charTrees[] = CharTree::fromString(strrev($string));
                 }
+                return CharTree::fromArray($charTrees);
             }
         );
 
@@ -66,10 +66,9 @@ final class DefaultFormatterTest extends TestCase
             '70,11',
             '11 ,70',
         ];
-
-        self::assertEquals($expected, array_unique(array_intersect(iterator_to_array($formatter->apply([
+        self::assertEquals($expected, array_intersect(iterator_to_array($formatter->apply([
             new DateTime('2001-02-03 07:11:19'),
             new DateTime('1907-11-19 21:13:58'),
-        ]), false), $expected)), '', 0, 10, true);
+        ]), false), $expected), '', 0, 10, true);
     }
 }